+++ src/factory/Cargo.toml
@@ -11,10 +11,10 @@ no_api = []
 candid = "0.7"
 ic-cdk = "0.5"
 ic-cdk-macros = "0.5"
-ic-canister = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-canister", tag = "v0.3.9" }
-ic-helpers = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-helpers", tag = "v0.3.9" }
-ic-storage = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-storage", tag = "v0.3.9" }
-ic-factory = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-factory", tag = "v0.3.9" }
+ic-canister = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-canister", branch = "awkure/pre-update-hook" }
+ic-helpers = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-helpers", branch = "awkure/pre-update-hook" }
+ic-storage = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-storage", branch = "awkure/pre-update-hook" }
+ic-factory = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-factory", branch = "awkure/pre-update-hook" }
 serde = "1.0"
 thiserror = "1.0"
 token = { path = "../token", features = ["no_api"] }
