+++ src/token/Cargo.toml
@@ -14,9 +14,9 @@ ic-cdk = "0.5"
 ic-cdk-macros = "0.5"
 num-traits = "0.2"
 serde = "1.0"
-ic-canister = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-canister", tag = "v0.3.9" }
-ic-helpers = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-helpers", tag = "v0.3.9" }
-ic-storage = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-storage", tag = "v0.3.9" }
+ic-canister = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-canister", branch = "awkure/pre-update-hook" }
+ic-helpers = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-helpers", branch = "awkure/pre-update-hook" }
+ic-storage = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-storage", branch = "awkure/pre-update-hook" }
 ic-certified-map = "0.3"
 serde_cbor = "0.11"
