+++ src/token/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod canister;
-pub mod core;
+pub mod exports;
 pub mod ledger;
 pub mod principal;
 pub mod state;
