+++ src/token/src/main.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 
 mod canister;
-mod core;
+mod exports;
 mod ledger;
 mod principal;
 mod state;
@@ -13,9 +13,9 @@ fn main() {}
 #[cfg(not(any(target_arch = "wasm32", test)))]
 fn main() {
     use token::canister::TokenCanister;
-    use token::core::ISTokenCanister;
+    use token::exports::TokenCanisterExports;
 
-    let trait_idl = <TokenCanister as ISTokenCanister>::get_idl();
+    let trait_idl = <TokenCanisterExports as TokenCanister>::get_idl();
 
     let result = candid::bindings::candid::compile(&trait_idl.env.env, &Some(trait_idl.actor));
     print!("{result}");
